Added more functionality to calculator
The behavior of the main loop is now to parse commands, and if
none match call calculate(). Commands are:
	print - prints the stack
	clear - clears the stack
	quit  - exits the application
calculate() now takes the input string and the stack, which is
persistant in the scope of main().
This new behavior means that the final test in test_rpn.py does
not test for the correct behavior (calculate doesn't care about
the stack size), so I removed it.