Changed the interface of calculate() and some tests
test_bad_input now tests invalid operators, invalid stack states
(too few operands) and makes sure that stack state is preserved
in the event of an error.

This has all been implemented by adding another error check in
calculate() for IndexError, which can (for now) only occur when
popping from an empty stack, and only returning the modified stack
if the command sequence was valid. I would prefer to handle this
via another method but the reference semantics of python are
alluding me at this juncture.